Change weight type to float

+++ src/TransitionStrategy.java
@@ -1,4 +1,4 @@
 public interface TransitionStrategy {
 
-    double getPrice(int weight);
+    double getPrice(float weight);
 }
